requirements.txt
asgiref==3.2.7
astroid==2.4.2
attrs==19.3.0
awsebcli==3.18.1
bcrypt==3.1.7
beautifulsoup4==4.9.1
blessed==1.17.9
botocore==1.15.49
CacheControl==0.12.6
cached-property==1.5.1
cachetools==4.1.1
cement==2.8.2
certifi==2020.4.5.1
cffi==1.14.1
chardet==3.0.4
colorama==0.4.3
cryptography==3.0
defusedxml==0.6.0
dj-database-url==0.5.0
Django==3.0.7
django-allauth==0.41.0
django-bootstrap-datepicker-plus==3.0.5
django-bootstrap4==1.1.1
django-cors-headers==3.4.0
django-crispy-forms==1.9.1
django-debug-toolbar==2.2
django-heroku==0.3.1
django-rest-auth==0.9.5
django-tinymce==3.0.2
django-tinymce4-lite==1.8.0
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
docker==4.2.2
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
docutils==0.15.2
future==0.16.0
gcloud==0.17.0
googleapis-common-protos==1.52.0
grpcio==1.30.0
gunicorn==20.0.4
httplib2==0.18.1
idna==2.7
isort==4.3.21
jmespath==0.10.0
jsmin==2.2.2
jsonschema==3.2.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
msgpack==1.0.0
oauth2client==3.0.0
oauthlib==3.1.0
paramiko==2.7.1
pathspec==0.5.9
Pillow==7.1.2
pip-autoremove==0.9.1
protobuf==3.12.2
psycopg2==2.8.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.5.3
PyNaCl==1.4.0
pyrsistent==0.16.0
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.3.1
requests==2.20.1
requests-oauthlib==1.3.0
rsa==4.6
semantic-version==2.5.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
termcolor==1.1.0
texttable==1.6.2
toml==0.10.1
uritemplate==3.0.1
urllib3==1.24.3
wcwidth==0.1.9
websocket-client==0.57.0
whitenoise==5.0.1
wrapt==1.12.1
